rename: move stories on stories dir
