+++ src/stories/checkbox.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Checkbox } from './checkbox';
+import { Checkbox } from '@/features/core';
 
 const meta = {
   component: Checkbox,
