+++ src/stories/label.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Label } from './label';
+import { Label } from '@/features/core';
 
 const meta = {
   component: Label,
